Add BlockNumber index to avoid lock wait timeouts.
The periodic drop history job searches for the first block after a specific timestamp. This is done in a transaction after some deletions, actually locking the table from other concurrent inserts. The transactions table has no index on BlockNumber used in the OrderBy clause, and MySQL is not able to use the Timestamp index on the Where clause to optimize the query, resulting in un-needed long table locks that sometimes exceed the MySQL lock wait timeout, causing fatal errors and indexer shutdown.